+++ src/Piece/hooks/index.js
@@ -1,2 +1,3 @@
+export { usePieceHoldingAbandonmentCheck } from './usePieceHoldingAbandonmentCheck';
 export { usePieceQuickReceiving } from './usePieceQuickReceiving';
 export { usePieceStatusChangeLog } from './usePieceStatusChangeLog';