+++ src/Piece/hooks/usePieceHoldingAbandonmentCheck/index.js
@@ -1,0 +1,1 @@
+export { usePieceHoldingAbandonmentCheck } from './usePieceHoldingAbandonmentCheck';