+++ src/ReceivingList/ExportSettingsModal/utils/fetchExportResources.test.js
@@ -1,6 +1,7 @@
 import {
   CONTRIBUTOR_NAME_TYPES_API,
   fetchExportDataByIds,
+  HOLDINGS_API,
   IDENTIFIER_TYPES_API,
   ITEMS_API,
   LINES_API,
@@ -9,7 +10,6 @@ import {
   VENDORS_API,
 } from '@folio/stripes-acq-components';
 
-import { HOLDINGS_API } from '../../../common/constants';
 import {
   fetchContributorNameTypesExportData,
   fetchIdentifierTypesExportData,